Add dynamic row heights based on font used. (Closes #3)
Deprecate 'title_length', ignore it for now. It is no longer needed with the update to Adafruit_CircuitPython_Display_Text removing 'max_glyphs'.

Refactor the scaling so that the 'text_scale' is not a multiple of the 'title_scale'.

Correctly add multiple rows with the color-cycling effect. Previously all new rows had the same color.